REFACTOR/ADD move noise gen code to butterfly.py, rescale relative to init stds when normalize_perturb is set
REFACTOR only allow noise to be generated at perturb time as this simplifies code
ADD rescaling means perturb_scale is proportional to init (i.e. perturb_scale=1 means same magnitude as random init)
ADD log per layer l2 stats to check for distribution shifts of batch noise over training time
TEST check that computed init stds match actual initializations

+++ lmc/utils/setup_training.py
@@ -235,7 +235,7 @@ def configure_lr_scheduler(
     optimizer: optim.Optimizer,
     training_steps: int,
     lr_scheduler: str = None,
-    warmup_ratio: int = 0,
+    warmup_ratio: float = 0,
     lr_schedule: dict = None,
     global_step: int = 0,
     warmup_steps: int = None,
